archive: mark EMetaBlob denc incompatible release before 60730271d4
in
https://github.com/ceph/ceph/commit/60730271d49a3c16bee6ff60c402b924a0280efc,
the encoding schema version of EMetaBlob bumped from 4 to 5. the older
release containing this change was v0.58

in
https://github.com/ceph/ceph/commit/d62c67ca2041797fc7aa212ae4a37ba91e5620e7,
the struct_v of EMetaBlob::fullbit::encode was bumped from 7 to 8, , the
oldest release containing this change was hammer rc1, i.e., v0.93.

in
https://github.com/ceph/ceph/pull/35931/commits/0127f2b01928b515bbbb7811976012ec97d3a957,
the backward compatible code of EMetaBlob::fullbit was dropped, so it
does not decode any EMetaBlob::fullbit whose struct_v belongs to [2, 8)
anymore.

in this change, all EMetaBlob* corpus created before 9.3 or 0.93 are
marked forward incompatible, so the denc tests with them are skipped if
the dencoder version is higher than those versions.

this change helps us to drop the backward compatible code to decode
EMetaBlob. as it is not neeed anymore after 2 releases.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>